refactor: use a provider for loading
